oblivion/requirements.txt
pandas==2.0.2
discord.py==2.3.2
python-dotenv==0.21.0